requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
requests
aiohttp
